requirement.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-ninja==0.21.0
idna==3.4
psycopg2==2.9.6
pydantic==1.10.7
requests==2.29.0
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
